+++ app/overrides/remove_welcome_latest_projects.rb
@@ -1,5 +1,0 @@
-Deface::Override.new virtual_path: 'welcome/index',
-                     name:         'remove-welcome-latest_projects',
-                     replace:      'div.projects',
-                     original:     'bd5d94c8882c7f599059c917195497b10d519b64',
-                     partial:      'welcome/remove_projects'