+++ CHANGELOG.rst
@@ -3,8 +3,12 @@ Changelog
 
 0.5.9-dev
 +++++++++
+
 - user group support for issue auto assign
 - optimize deface overwrite path
+- drop remove latest projects support (because Redmine 3.2 has dropped latest projects)
+- add permission for log time on closed issues - make sure you adjust our permissions!
+- code cleanups and bug fixes
 
 0.5.8
 +++++